Update HITRAN module to use class-based query API
fixes #824
