+++ docs/hitran/hitran.rst
@@ -7,26 +7,29 @@ HITRAN Queries (`astroquery.hitran`)
 Getting started
 ===============
 
-This module provides an interface to the `HITRAN`_ database API.  It can
-download a data file including transitions for a particular molecule in a given
-wavenumber range.  The file is downloaded in the default cache directory 
-``~/.astropy/cache/astroquery/hitran`` and can be opened with a reader function
-that returns a table of spectral lines including all accessible parameters.
+This module provides an interface to the high-resolution transmission molecular
+absorption database API (`HITRAN`_).  The current version of the database
+contains a compilation of spectroscopic parameters for 49 molecular species
+along with their most significant isotopologues.  Using the `hitran` module you
+can search transitions for a particular molecule in a given wavenumber range.
 
 Examples
 ========
 
 This will download all transitions of the main isotopologue of water between
-the wavenumbers of 3400 and 4100 cm\ :sup:`-1`\ .
+the wavenumbers of 3400 and 4100 cm\ :sup:`-1`\ .  The expected type for the
+parameters ``min_frequency`` and ``max_frequency`` is an AstroPy quantity.
+The data are returned as an `~astropy.table.Table` instance.
+
 
 .. code-block:: python
 
-    >>> import os
-    >>> from astroquery.hitran import read_hitran_file, cache_location, download_hitran
-    >>> download_hitran(1, 1, 3400, 4100)
-    >>> tbl = read_hitran_file(os.path.join(cache_location, 'H2O.data'))
-
-Transitions are returned as an `~astropy.table.Table` instance.
+    >>> from astropy import units as u
+    >>> from astroquery.hitran import Hitran
+    >>> tbl = Hitran.query_lines_async(molecule_number=1,
+                                       isotopologue_number=1,
+                                       min_frequency=0. / u.cm,
+                                       max_frequency=10. / u.cm)
 
 Reference/API
 =============
